@@ -6,28 +6,28 @@
   "author": "Ian Spence",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/bootstrap": "^5.1.10",
+    "@types/bootstrap": "^5.1.12",
     "@types/prop-types": "^15.7.5",
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.3",
+    "@types/react": "^18.0.10",
+    "@types/react-dom": "^18.0.5",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.22.0",
-    "@typescript-eslint/parser": "^5.22.0",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "^5.27.0",
     "autoprefixer": "^10.4.7",
-    "copy-webpack-plugin": "^10.2.4",
+    "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
-    "eslint": "^8.14.0",
-    "eslint-plugin-react": "^7.29.4",
+    "eslint": "^8.17.0",
+    "eslint-plugin-react": "^7.30.0",
     "eslint-webpack-plugin": "^3.1.1",
     "html-loader": "^3.1.0",
     "html-webpack-plugin": "^5.5.0",
-    "sass": "^1.51.0",
-    "sass-loader": "^12.6.0",
+    "sass": "^1.52.2",
+    "sass-loader": "^13.0.0",
     "source-map-loader": "^3.0.1",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.3.0",
-    "typescript": "^4.6.4",
-    "webpack": "^5.72.0",
+    "typescript": "^4.7.3",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
     "webpack-merge": "^5.8.0"
   },
@@ -41,7 +41,7 @@
     "@types/debounce-promise": "^3.1.4",
     "bootstrap": "^5.1.3",
     "debounce-promise": "^3.1.2",
-    "react-router-dom": "^5.3.1",
+    "react-router-dom": "^6.3.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0"
   }
