@@ -4,7 +4,7 @@ const ESLintPlugin = require('eslint-webpack-plugin');
 
 module.exports = {
     resolve: {
-        extensions: [".ts", ".tsx", ".js", ".html"]
+        extensions: ['.ts', '.tsx', '.js', '.html']
     },
 
     plugins: [
@@ -18,7 +18,7 @@ module.exports = {
             ]
         }),
         new ESLintPlugin({
-            extensions: [".ts", ".tsx"]
+            extensions: ['.ts', '.tsx']
         }),
     ],
 
@@ -29,19 +29,18 @@ module.exports = {
                 exclude: /node_modules/,
                 use: [
                     {
-                        loader: "ts-loader"
+                        loader: 'ts-loader'
                     }
                 ]
             },
-            // All output '.js' files will have any sourcemaps re-processed by 'source-map-loader'.
-            {
-                enforce: "pre",
-                test: /\.js$/,
-                loader: "source-map-loader"
-            },
             {
                 test: /\.(woff|woff2)$/,
-                type: 'asset/inline',
+                type: 'asset/inline'
+            },
+            {
+                enforce: 'pre',
+                test: /\.js$/,
+                loader: 'source-map-loader'
             },
             {
                 test: /\.s[ac]ss$/i,
@@ -61,16 +60,6 @@ module.exports = {
                 ]
             },
         ]
-    },
-
-    // When importing a module whose path matches one of the following, just
-    // assume a corresponding global variable exists and use that instead.
-    // This is important because it allows us to avoid bundling all of our
-    // dependencies, which allows browsers to cache those libraries between builds.
-    externals: {
-        'react': 'React',
-        'react-dom': 'ReactDOM',
-        'react-router-dom': 'ReactRouterDOM',
     },
 
     output: {
